Restructured tests to be scoped by module tested
